update allocations to use perms and filter

+++ app/Http/Controllers/AllocationController.php
@@ -7,9 +7,11 @@ use App\Models\Resource;
 use App\Models\Project;
 use App\Models\Demand;
 use App\Models\ResourceType;
+
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use App\Http\Requests\AllocationRequest;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Redirect;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\View\View;
@@ -35,11 +37,6 @@ class AllocationController extends Controller
     public function __construct(CacheService $cacheService)
     {
         $this->cacheService = $cacheService;
-
-        // $this->middleware('allocation:view', ['only' => ['index']]);
-        // $this->middleware('allocation:create', ['only' => ['create','store']]);
-        // $this->middleware('allocation:update', ['only' => ['update','edit']]);
-        // $this->middleware('allocation:delete', ['only' => ['destroy']]);
     }
 
     /**
@@ -47,6 +44,10 @@ class AllocationController extends Controller
      */
     public function index(Request $request): View
     {
+        $user = auth()->user();
+        if (!$user->can('allocations.index')) {
+            return Redirect::route('home')->with('warning', 'You do not have the necessary permissions to view the allocations page.');
+        }
         // Build our next twelve month array
         $nextTwelveMonths = [];
 
@@ -64,10 +65,7 @@ class AllocationController extends Controller
         $endDate = Carbon::now()->addYear()->startOfMonth();
 
         // Collect our resources who have a current contract
-        $resources = Resource::whereHas('contracts', function ($query) {
-            $query->where('start_date', '<=', now())
-                ->where('end_date', '>=', now());
-        })->with('contracts')->paginate();
+        $resources = $this->ResourceList();
 
         // Modify resource names to add [c] if the resource is not permanent
         foreach ($resources as $resource) {
@@ -92,6 +90,9 @@ class AllocationController extends Controller
      */
     public function create(): View
     {
+        if (!$user->can('allocations.create')) {
+            return Redirect::route('allocations.index')->with('warning', 'You do not have the necessary permissions to view the allocations page.');
+        }
         $allocation = new Allocation();
         $resources = Resource::all(); // Retrieve all resources
         $projects = Project::all(); // Retrieve all projects
@@ -115,6 +116,9 @@ class AllocationController extends Controller
      */
     public function show($id): View
     {
+        if (!$user->can('allocations.show')) {
+            return Redirect::route('allocations.index')->with('warning', 'You do not have the necessary permissions to view the allocations page.');
+        }
         $allocation = Allocation::find($id);
 
         return view('allocation.show', compact('allocation'));
@@ -186,10 +190,47 @@ class AllocationController extends Controller
 
     public function destroy($id): RedirectResponse
     {
+        if (!$user->can('allocations.delete')) {
+            return Redirect::route('allocations.index')->with('warning', 'You do not have the necessary permissions to view the allocations page.');
+        }
         Allocation::find($id)->delete();
 
         return Redirect::route('allocations.index')
             ->with('success', 'Allocation deleted successfully');
     }
 
+    private function ResourceList()
+    {
+        //get user 
+        $user = Auth::user();
+        // Check if the user is an owner of a team
+        if ($user->ownedTeams()->count() > 0) {
+            // Get the team's resources
+            Log::info("User is an owner of a team with resource type: " . $user->ownedTeams()->first()->resource_type);
+            $resource_type = ResourceType::where('name', $user->ownedTeams()->first()->resource_type)->first()->id;
+            Log::info("resource type: " . json_encode($resource_type));
+            $resources = Resource::whereHas('contracts', function ($query) {
+                $query->where('start_date', '<=', now())
+                    ->where('end_date', '>=', now());
+            })
+                ->where('resource_type', $resource_type)
+                ->with('contracts')->paginate();
+            Log::info("resources: " . json_encode($resources));
+        } elseif ($user->reportees->count() > 0) {
+            // check if the user is a manager
+            Log::info("User is a manager");
+            $reportees = $user->reportees;
+            //for each linked resource contract, check if the start date is before now and the end date is after now
+            $resourceIDs = $user->reportees->pluck('resource.id')->toArray();
+            $resources = Resource::whereIn('id', $resourceIDs)->whereHas('contracts', function ($query) {
+                $query->where('start_date', '<=', now())
+                    ->where('end_date', '>=', now());
+            })->with('contracts')->paginate();
+        } else {
+            Log::info("User is not an owner of a team or a manager");
+            // otherwise just return the user's resource
+            $resources = Resource::where('id', $user->resource_id)->with('contracts')->paginate();
+        }
+        return $resources;
+    }
 }
